[1767] Refactor tags renderer service to be a component instead
