+++ app/components/api/guidance/release_notes/tags_component.html.erb
@@ -1,0 +1,1 @@
+<%= all_tags %>
