+++ app/views/api/guidance/release_notes.html.erb
@@ -5,8 +5,6 @@
     <span class="govuk-caption-m"><%= release_note.date %></span>
     <%= release_note.title %>
   </h2>
-  <div class="release-notes-tags">
-    <%= release_note.tags  %>
-  </div>
+  <%= render API::Guidance::ReleaseNotes::TagsComponent.new(tags: release_note.tags) %>
   <%== release_note.body %>
 <% end %>
